requirements.txt
Flask==2.0.1
Pillow==9.5.0
gradio-client==0.6.1
gunicorn==20.1.0
Werkzeug==2.2.3 